Add map example for skidpad and straights
MS Painted some straights and figure eights (that rhymes!) Note that these are NOT intended to be made to the precise scale of the competition in real life. IMO, any algorithm that manages accel or skidpad at one scale should be able to generalize to other scales. Additionally such a planning algorithm should be able to handle.. imprecision in track construction